Refactoring according to review comments

+++ src/chainparams.cpp
@@ -406,7 +406,7 @@ public:
 
         adminParams.m_blockToAdminKeys.emplace(0, CreateRegTestAdminKeys());
 
-        snapshotParams.createSnapshotPerEpoch = gArgs.GetArg("-createsnapshot", 1);
+        snapshotParams.createSnapshotPerEpoch = static_cast<uint16_t>(gArgs.GetArg("-createsnapshot", 1));
 
         base58Prefixes[PUBKEY_ADDRESS] = std::vector<unsigned char>(1,111);
         base58Prefixes[SCRIPT_ADDRESS] = std::vector<unsigned char>(1,196);
